move some asset modal molecule func into hooks
